\name{DaubechiesSim-package}
\alias{DaubechiesSim-package}
\alias{DaubechiesSim}
\docType{package}
\title{\packageTitle{DaubechiesSim}}
\description{\packageDescription{DaubechiesSim}}
\section{Package Content}{\packageIndices{DaubechiesSim}}
\author{\packageAuthor{DaubechiesSim}}
\section{Maintainer}{\packageMaintainer{DaubechiesSim}}
\keyword{package}
